批量修改价格

+++ voyageone-service2/voyageone-service2-dao/src/main/java/com/voyageone/service/daoext/cms/CmsBtJmPromotionSkuDaoExt.java
@@ -31,9 +31,9 @@ public interface CmsBtJmPromotionSkuDaoExt {
 
     List<Map<String, Object>> selectExportListByPromotionId(int promotionId);
    //更新的dealPrice大于market_price的记录
-    CmsBtJmPromotionSkuModel  selectNotUpdateDealPrice(@Param("listPromotionProductId") List<Long> listPromotionProductId, @Param("dealPrice") String dealPrice);
+    CmsBtJmPromotionSkuModel  selectNotUpdateDealPrice(@Param("listPromotionProductId") List<Integer> listPromotionProductId, @Param("dealPrice") String dealPrice);
 
-    int batchUpdateDealPrice(@Param("listPromotionProductId") List<Long> listPromotionProductId, @Param("dealPrice") String dealPrice);
+    int batchUpdateDealPrice(@Param("listPromotionProductId") List<Integer> listPromotionProductId, @Param("dealPrice") String dealPrice);
 
     List<SkuPriceBean> selectJmSkuPriceInfoListByPromotionProductId(int promotionProductId);
