+++ voyageone-service2/voyageone-service2-dao/src/main/java/com/voyageone/service/daoext/cms/CmsBtPromotionCodesDaoExtCamel.java
@@ -18,7 +18,7 @@ import java.util.Map;
 public interface CmsBtPromotionCodesDaoExtCamel {
     int deleteByPromotionId(int promotionId);
     int deleteByPromotionCodeList(Map<String,Object> map);
-    int updateJmPromotionPrice(@Param("jmPromotionId")int jmPromotionId, @Param("listPromotionProductId") List<Long> listPromotionProductId);
+    int updateJmPromotionPrice(@Param("jmPromotionId")int jmPromotionId, @Param("listPromotionProductId") List<Integer> listPromotionProductId);
 
     /**
      * 取得当前有效的活动下的所有产品
